ui/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
httpx
google-cloud-bigquery
db-dtypes
